Kelly Broome epic

+++ scripts/deploy.sh
@@ -8,44 +8,52 @@ dest="${url}:~/"
 
 echo This script deploys the EventEnsure app to the ec2 instance located at ${url}
 
-cd ${base}
-echo Changed directory to $(pwd)
+echo "Are you certain you want to deploy"
+read certain
+if [ ${certain} == 'y' ]
+then
+    cd ${base}
+    echo Changed directory to $(pwd)
 
-npm run build
-echo npm has run the build script located in package.json
+    npm run build
+    echo npm has run the build script located in package.json
 
-scp -i ${pem} server.js package.json ${dest}new/
-scp -i ${pem} -r backend/ ${dest}new/
-scp -i ${pem} -r public/ ${dest}new/
-echo Files were sent to the ec2 instance
+    scp -i ${pem} server.js package.json ${dest}new/
+    scp -i ${pem} -r backend/ ${dest}new/
+    scp -i ${pem} -r public/ ${dest}new/
+    echo Files were sent to the ec2 instance
 
-ssh -i ${pem} ${url} << END
-    echo ssh was successful
+    ssh -i ${pem} ${url} <<END
+        echo ssh was successful
 
-    pm2 kill
-    echo current instance of the website was killed
+        pm2 kill
+        echo current instance of the website was killed
 
-    rm -rf server/backend
-    rm -rf server/public/
-    rm server/package.json
-    rm server/server.js
-    echo old version of the website was removed
+        rm -rf server/backend
+        rm -rf server/public/
+        rm server/package.json
+        rm server/server.js
+        echo old version of the website was removed
 
-    mv new/backend/ server/backend/
-    mv new/server.js server/
-    mv new/package.json server/
-    mv new/public server/
-    echo files have been moved to teh appropriate places
-    echo stage is set for the next deploy
+        mv new/backend/ server/backend/
+        mv new/server.js server/
+        mv new/package.json server/
+        mv new/public server/
+        echo files have been moved to teh appropriate places
+        echo stage is set for the next deploy
 
-    cd server/
-    npm install --only=production
-    echo npm install was run
+        cd server/
+        npm install --only=production
+        echo npm install was run
 
-    echo starting new instance of the website
-    ./../init.sh
+        echo starting new instance of the website
+        ./../init.sh
 
-    exit
+        exit
 END
-duration=$SECONDS
-echo deployment took "$(($duration / 60)) minutes and $(($duration % 60)) seconds"
+
+    duration=$SECONDS
+    echo deployment took "$(($duration / 60)) minutes and $(($duration % 60)) seconds"
+else
+    echo cancelling
+fi
